+++ NEWS.md
@@ -6,7 +6,8 @@ BUGFIXES :
 * `.getSimOptions()` reads and returns the new converted study version format (ex : 9.0 => 900) in output mode
 * `.importOutput()` Fixed area outputs import for Antares ≥ 9.3 when output columns differ between areas.
 * `.readDistrictsDefAPI()` treats case when `apply-filter` property is not written.
-.
+* `.getStartDate()` is case insensitive to treat property `first-month-in-year` 
+
 # antaresRead 2.9.3
 
 NEW FEATURES (cf. Antares v9.3 changelog) :